+++ crates/smartvaults-desktop/src/constants.rs
@@ -5,11 +5,11 @@ pub const APP_NAME: &str = "Smart Vaults";
 pub const APP_LOGO: &[u8] = include_bytes!("../static/img/smartvaults.svg");
 pub const APP_DESCRIPTION: &str = env!("CARGO_PKG_DESCRIPTION");
 
-pub const BIGGER_FONT_SIZE: u8 = 19;
-pub const BIG_FONT_SIZE: u8 = 17;
-pub const DEFAULT_FONT_SIZE: u8 = 15;
-pub const SMALL_FONT_SIZE: u8 = 13;
-pub const SMALLER_FONT_SIZE: u8 = 11;
+pub const BIGGER_FONT_SIZE: u16 = 19;
+pub const BIG_FONT_SIZE: u16 = 17;
+pub const DEFAULT_FONT_SIZE: u16 = 15;
+pub const SMALL_FONT_SIZE: u16 = 13;
+pub const SMALLER_FONT_SIZE: u16 = 11;
 
-pub const DEFAULT_ICON_SIZE: u8 = 20;
-pub const BIG_ICON_SIZE: u8 = 22;
+pub const DEFAULT_ICON_SIZE: u16 = 20;
+pub const BIG_ICON_SIZE: u16 = 22;
